requirements.txt
ansible
python-dracclient
python-ironicclient
